+++ model/sub-band.h
@@ -127,8 +127,8 @@ class SubBand : public Object
     double GetMaxTxPowerDbm() const;
 
   private:
-    double m_firstFrequencyHz;   //!< Starting frequency of the subband, in Hz
-    double m_lastFrequencyHz;    //!< Ending frequency of the subband, in Hz
+    uint32_t m_firstFrequencyHz; //!< Starting frequency of the subband, in Hz
+    uint32_t m_lastFrequencyHz;  //!< Ending frequency of the subband, in Hz
     double m_dutyCycle;          //!< The duty cycle that needs to be enforced on this subband
     Time m_nextTransmissionTime; //!< The next time a transmission will be allowed in this subband
     double m_maxTxPowerDbm; //!< The maximum transmission power that is admitted on this subband
